changed config
