+++ src/app/api/highscores/route.js
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { connectToDatabase } from "@/lib/mongodb";
+import { connectToDatabase } from "@/lib/mongod.mjs";
 
 export async function GET() {
   try {
